third_party/blink/web_tests/external/wpt/css/css-values/container-progress-serialize.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] 'calc(container-progress(width, 0px, 1px) / 1000)' as a specified value should serialize as 'calc(0.001 * container-progress(width, 0px, 1px))'.
  assert_equals: 'calc(container-progress(width, 0px, 1px) / 1000)' and 'calc(0.001 * container-progress(width, 0px, 1px))' should serialize the same in specified values. expected "calc(0.001 * container-progress(width, 0px, 1px))" but got "calc(container-progress(width, 0px, 1px) / 1000)"
[FAIL] 'scale(calc(container-progress(width, 0px, 1px) / 1000))' as a specified value should serialize as 'scale(calc(0.001 * container-progress(width, 0px, 1px)))'.
  assert_equals: 'scale(calc(container-progress(width, 0px, 1px) / 1000))' and 'scale(calc(0.001 * container-progress(width, 0px, 1px)))' should serialize the same in specified values. expected "scale(calc(0.001 * container-progress(width, 0px, 1px)))" but got "scale(calc(container-progress(width, 0px, 1px) / 1000))"
Harness: the test ran to completion.
